+++ src/ytextractor.nim
@@ -1,6 +1,6 @@
 #[
   Created at: 08/03/2021 19:58:57 Tuesday
-  Modified at: 08/20/2021 12:48:20 PM Friday
+  Modified at: 09/10/2021 11:59:23 PM Friday
 ]#
 
 ##[
@@ -13,16 +13,19 @@ import ytextractor/channel; export channel
 when isMainModule and not defined js:
   # debug purposes
   from std/times import `$`
-  import ytextractor/exports
-  import std/json
+  # import ytextractor/exports
+  # import std/json
 
-  # var vid = initYoutubeVideo "jjEQ-yKVPMg".videoId
-  # discard vid.update()
-  # echo vid
+  var vid = initYoutubeVideo "jjEQ-yKVPMg".videoId
+  discard vid.update()
+
+  echo "Video data:"
+  echo vid
 
   # echo extractVideo("https://www.youtube.com/watch?v=u8ZP9g-RKA8")
 
   var chan = initYoutubeChannel "https://www.youtube.com/channel/UC3aGq0eFrvrjM4F1dLUo87A".channelId
-  # var chan = initYoutubeChannel "https://www.youtube.com/channel/UCsBjURrPoezykLs9EqgamOA".channelId
   discard chan.update(home)
+
+  echo "\nChannel data:"
   echo chan