+++ src/ytextractor/channel.nim
@@ -1,6 +1,6 @@
   #[
   Created at: 08/09/2021 12:10:05 Monday
-  Modified at: 08/20/2021 01:04:39 PM Friday
+  Modified at: 09/10/2021 11:43:48 PM Friday
 ]#
 
 ##[
@@ -13,7 +13,7 @@
 {.experimental: "codeReordering".}
 
 from std/times import DateTime, Duration, initDuration, now
-from std/json import JsonNode, items, hasKey, `{}`, getStr, getInt, getBool
+from std/json import JsonNode, items, hasKey, `{}`, getStr, getInt, getBool, len
 from std/strformat import fmt
 from std/strutils import parseInt, multiReplace, find, strip, contains, parseEnum
 from std/uri import parseUri